// apps/electron/src/renderer/src/gen/api/schemas/post.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Techoo API
 * API for the Techoo task management application with OpenAPI documentation
 * OpenAPI spec version: 1.0.0
 */
import type { PostEventsItem } from './postEventsItem'
import type { PostTodosItem } from './postTodosItem'

export interface Post {
  id: string
  body: string
  /** @minimum 0 */
  posted_at: number
  events: PostEventsItem[]
  todos: PostTodosItem[]
}